requirements.txt
python-fasthtml
uvicorn>=0.29
sqlite_utils
requests